Adds review changes and merges dark mode PR
