+++ src/components/overview/types.ts
@@ -1,0 +1,15 @@
+import { TeamIntervalsOverviewDto } from "@/app/api/teams/overview/intervals/route";
+
+export interface TeamIntervalsOverviewProps {
+  teamIntervalsOverview: TeamIntervalsOverviewDto[];
+}
+
+export interface TeamIntervalsChartData {
+  start: number;
+  end: number;
+  teamScores: { name: string; score: number }[];
+}
+
+export interface TeamUtils {
+  teamNameToIndex: Map<string, number>;
+}
