+++ src/components/overview/utils.ts
@@ -1,0 +1,19 @@
+export function indexToColor(index: number): string {
+  //switch case for index
+  switch (index) {
+    case 0:
+      return "#adfa1d";
+    case 1:
+      return "#A4FFFF";
+    case 2:
+      return "#FFB86C";
+    case 3:
+      return "#242c54";
+    case 4:
+      return "#A4FFFF";
+    case 5:
+      return "#FFFFFF";
+    default:
+      return "#adfa1d";
+  }
+}
